VSVGVQ-80 move createRegistrationForUser method to accountviewcontroller
